+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./ids";
+export * from "./types";
 export * from "./builder";
